core: Mark our limits
We won't be able to run more than a few less external apps than
ulimit -u, or ulimit -n / 2, so it's good if we know when we'll
hit that limit so we can avoid sending more jobs to our workers
when we do. This patch at least lets us become aware of our own
limites.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2376 5f96b256-904b-4d8d-8c98-d829582c6739