core: Make nagios work with path-located binary again
When we're calling nagios as just plain vanilla "nagios", the path
canonicalization (lots of points on scrabble for that one) fails
pretty hard. We take the easy way out and ignore the path completely
if it's path-located. Since we're using execvp() to spawn our workers
anyway we needn't bother with it.

As a neat bonus, this moves the nagios_binary_path checking from
the completely braindead location where I first put it.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2393 5f96b256-904b-4d8d-8c98-d829582c6739