core: fflush() logfile after every write
Logfiles are used to generate reports, so it's kind of important
that they're up to date. fflush() isn't so bad as closing and
reopening the file constantly, as the kernel doesn't have to
flush it to disk every time but can handle it in its buffers.
If we end up in a kernel panic, we will lose *some* data, but
those are pretty rare on production systems so we'll take our
chances.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2396 5f96b256-904b-4d8d-8c98-d829582c6739