Stop mentioning and using check_reaper_interval
It is nearly useless for us since we don't reap checks on specific
times, but rather wait for our workers to feed us the results as
they become available, meaning that the check_reaper_interval is
all but insignificant for us.

Instead we take a semi-wild guess on how long checks will take to
run on average, guess (again) that each check will require about
0.02 seconds of cpu time (pessimistic, I know, but better safe than
sorry) and then give out some advice based on that.

While at it, we also check the nproc_limit and nofile_limits and
make sure they will allow us to run the number of checks we must
be able to execute in parallel.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2401 5f96b256-904b-4d8d-8c98-d829582c6739