lib/iocache: Never iocache_reset() the buffer in iocache_available()
Probably the most common stream protocols in the world use a header
which determines the size of the following body, but it breaks down
if the check for available data suddenly resets the buffer from under
our feet. Consider this scenario:

hdr = (struct hdr *)iocache_use_size(ioc, sizeof(*hdr));
if (hdr->body_len > iocache_available(ioc))
	iocache_unuse_size(ioc, sizeof(*hdr));

The iocache_use_size() call may deplete all available data in the
buffer, causing the following iocache_available() check to reset it,
which in turn causes the iocache_unuse_size() call to fail and the
stream to become unsynchronized. That's a Very Bad Thing(tm) indeed,
but this patch fixes it.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2402 5f96b256-904b-4d8d-8c98-d829582c6739