core: Don't let core-only files carry over to plugins
It's an unnecessary waste of resources that we can do without.
By using fcntl() to mark them for closure on execv() calls, the
kernel doesn't have to copy their filedescriptor structures and
we actually get away with a microscopic saving in cpu time used
as well, which is neat.

The larger benefit is that misbehaving plugins have no chance
at all to screw up our internal data.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2416 5f96b256-904b-4d8d-8c98-d829582c6739