lib/kvvec: Parse last key/value pair even if it lacks a pair separator
The idiom "foo=lala;bar=hoopla;toodles=banana" makes a lot of sense
for passing data between variables, but we used to miss the last
variable in the key/value vector library if it didn't have a pair
separator.

With this patch the last variable is included iff it's the very
last one *AND* we've already found the (non-optional) key/value
separator (the equal sign '=' in any sane application), which
makes the above idiom work as intended.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2419 5f96b256-904b-4d8d-8c98-d829582c6739