lib/worker: Kill off all children when we exit
In large (as in, really large) networks, we can easily leave
thousands of processes behind when we're about to reload. That's
pretty bad form and means a ton of processes will contend with
the new processes for cpu time, meaning we can get loadspikes in
the thousands. Not very nice.

Instead we consolidate the worker exiting stuff and make sure we
kill (and reap) our children before we exit properly from the
workers.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2421 5f96b256-904b-4d8d-8c98-d829582c6739