query-handler: Enforce the space character
We really don't want a bazillion different query handlers, and this
should discourage the one-query-only handlers from registering with
us. They still can, but queries to them will have to look like so:
  @somewhere<sp><nul>

which has a very shoddy feel about it and will hopefully cause an
infinite amount of shame and grief to fall on the House Of The
Offender.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2423 5f96b256-904b-4d8d-8c98-d829582c6739