core: Log incoming worker registration requests
In order to do this (easily), we must change the kv pair separator
so we can log the whole thing as a single line. We do this by using
';' instead of '\n' and log the raw request as and when it comes in.

While at it, we also add error logging for when registration fails
and plug a microscopic leak of sizeof(worker_process) when we fail
to parse the request.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2427 5f96b256-904b-4d8d-8c98-d829582c6739