core: Implement loadcontrol
This is currently the basis for experimentation to prevent loads
in the multiples of thousands when we're running configurations
too large for small systems. The current behaviour of pushing the
system load through the roof is simply unacceptable, prevents
proper system management and causes checks to time out due to the
extreme load (which in turn leads to even higher load).

Module authors should refrain from using the loadctl structure,
as it's likely to change in the next couple of days.

The hope is that this will cause latency to increase instead of
insane system load bringing the entire system down.

By default, nothing what so ever will change, except that we
now avoid asking workers to spawn jobs when we know that won't
work due to system ulimits.

Remote workers and caching special-purpose workers are not accounted
for yet and will need a patch to both can_spawn_jobs() and the worker
management code.

While at it, we fix the limits for how many external apps we can
run based on the number of filedescriptors we can have. Those are
per process, and not per user, so we need to use them with the
number of local workers we spawn. It's unlikely to be a limit
anyway, but if we're going to use it we might as well do it right.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2434 5f96b256-904b-4d8d-8c98-d829582c6739