core: Make queryhandler a bit easier to work with
Most 'set' operations should print an OK message on success, but
earlier they had to print that themselves. Stuff that already prints
a lot of info should return 0 to avoid the '200: OK' message to the
client.

While at it, we add support for a bunch more http-ish error codes,
making it trivial for handlers to get sensible feedback to clients.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2437 5f96b256-904b-4d8d-8c98-d829582c6739