core: Thinkos and typos in set_loadctl_options()
With the previous code, it was first of all impossible to set the
most important metric of all; jobs_limit.

Secondly, one could set a limit that was below the jobs_min, or
above jobs_max. We shan't tolerate such foolishness. Considering
how easy it is to change the load control parameters it's no
great chore to set the max and min parameters alongside the
current active limit.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2444 5f96b256-904b-4d8d-8c98-d829582c6739