lib/iobroker: Add output polling to the iobroker's repertoire
For some reason, I actually forgot about it, and then in my
hazed daze of tiredness realized that without it it would only
be an "iBroker", and before you know it we'd have to rename
Nagios to something feline and retarded because we happened to
prefix a noun with an I. Now that we've amended the prefix with
an O, and thus completed it, I feel better already!

It's also useful, since it means we can poll (efficiently) to
see which workers are ready to receive input, in case one of
them happens to be a slow reader, suffer from dyslexia or just
doesn't feel like reading when we want it to.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2446 5f96b256-904b-4d8d-8c98-d829582c6739