base/workers: Don't write to uninitialized memory
get_job_id returns -1 when there are no free slots in the worker. We
didn't handle this case with anything other than a comment - the end
result is that the nagios core will assign the -1 slot of the worker,
causing memory errors when we tried to write to the nonexistent slot.

This seems to be what sometimes created crashes when shutting
down/restarting the process. While I haven't been able to create a
completely reproducible test case, I have a fairly large, completely
unrelated test suite that used to cause a crash roughly 1/2 the time it
was executed, and this patched stopped that.

The FIXME to figure out somewhere else to put the check is still in its
place - I still don't do that - but not dumping core because we've got a
sizable workload seems reasonable. It's unlikely that another worker is
available to spread the workload for us, so the error returned about us
being too busy to do anything should normally already be quite
indirectly noticable anyway.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2450 5f96b256-904b-4d8d-8c98-d829582c6739