base/events: Don't get stuck in busy-wait loop when poll returns EINVAL
In a couple of instances, by completely breaking libnagios, I've managed
to create situations where there are no workers connected to nagios.
When that happens, nagios would get stuck in an infinite loop, where
polling for events always returns -1 (EINVAL) immediately, and nagios
would respond to that by immediately trying again, making my machine
warm and tired. Exiting the event loop when that happens seems more
reasonable - we can't run any more checks, and we can't notify anyone
about this, because we have no worker to do so for us.

It would be possible to set sigrestart to TRUE, to force nagios to
restart all its workers again, yet this could create the same crash loop
that I'm trying to fix, only somewhat larger and slower - if all sockets
died, something is seriously broken, so make it possible for an external
watchdog daemon to find out.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2451 5f96b256-904b-4d8d-8c98-d829582c6739