base/workers: Fix crash when freeing the last core worker
When freeing the last core worker, we'd set the worker list to NULL, yet
keep the list len at 1. This caused free_worker_memory to try to lookup
workers from NULL, creating a SEGFAULT.

By always decrementing len, even when we're destroying the last worker,
we'll actually realize that we're out of workers, leave NULL alone, and
shut down cleanly.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2452 5f96b256-904b-4d8d-8c98-d829582c6739