base/workers: Fix detection of check basename for special workers
strrchr may search from the end of the string, but it still expects the
pointer it is provided to point to the beginning of the string. Thus, we'd
only check if the last byte was a slash, rather than for the last slash.

Manipulating the cmd_name for basename discovery is unnecessary, so
don't.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2453 5f96b256-904b-4d8d-8c98-d829582c6739