Corrected logic to reinstate downtimes after Nagios restart.
Previously the flexible downtime start was never recorded on a
Nagios core shutdown and thus a flexible downtime could not be
restarted correctly after a core restart. This commit fixes
this issue by recording the flexible downtime start in the 
status.dat and retention.dat.

Also, there was an issue where some downtimes never ended without
a core restart and some downtimes never restared after a core 
restart. This is also fixed with this commit.


git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2459 5f96b256-904b-4d8d-8c98-d829582c6739