base/commands: Don't forget output in passive host checks
The passive host checks weren't assigning several of the variables that
it was supposed to, including output. As the output is quite interesting
in nagios checks, we should put that back in there.

While I'm at it, reorder and sync the assignments in passive host check
result with those in the passive service check result, so it's easier to
see if they're similar enough or not. This also makes it clear that we
were missing source and engine assignments.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2473 5f96b256-904b-4d8d-8c98-d829582c6739