xodtemplate: Fix obscure crash while loading configuration
We didn't initialize parent_map, meaning that it could have any value,
even though it appears it is not infrequently NULL, at least at first.

We then pointed it to a bitmap object iff
xodtemplate_servicedependency_list wasn't NULL, and finally freed it
unconditionally.

Thus, unless parent_map pointed at NULL or we have servicedependencies,
we'd get a crash.

Now, if we first add at least one servicedependency, run nagios, remove
all of them, and reload nagios, parent_map would still keep its old
value around, we wouldn't reallocate it, but we'd still free it, leading
to a crash.

This is prevented in the following ways:
 * First, make sure all bitmaps are NULL-initialized, clearly and
   obviously.
 * Second, only free parent_map if we did set it.
 * Third, set parent_map to NULL after we destroy() it - dangling
   pointers are evil.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2475 5f96b256-904b-4d8d-8c98-d829582c6739