lib/iobroker: Only close fds we're managing
In all places except one, iobroker makes sure it only touches the file
descriptors that's been registered with it. The last place is in the
destructor, which is called quite early in the nagios shutdown process,
making it impossible to log anything else after this.

This patch excludes any valid fds we're not supposed to do anything
with, meaning that, say, stdout won't get mercilessly closed.

In practice, without this patch, a restart would log any broker modules
being unloaded, while a shutdown wouldn't. With this patch, shutdowns do
log broker modules being unloaded.

Reported-by: Anton Löfgren <alofgren@op5.com>
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2476 5f96b256-904b-4d8d-8c98-d829582c6739