base/workers: Protect against allocation errors in the right places
In a recent patch we added a half-assed protection against memory
allocation errors by protecting against failure to allocate memory
for object check jobs.

This patch removes that particular check and replaces it with a
few others that are more general in nature and will prevent us
from accessing illegal memory no matter where or how allocation
misfires. It is an expanded derivative of a patch sent in by
Ricardo Maraschini, who drew my attention to the matter.

Hinted-at-by: Ricardo Maraschini <ricardo.maraschini@opservices.com.br>
Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2480 5f96b256-904b-4d8d-8c98-d829582c6739