query-handler: Do error logging when registering query handlers
It's retarded to have every module and in-core query handler have
to do its own error logging when we can do it in one place and
also include a legible error message in that one place instead of
attempting to guess at all the others.

While at it, we stop logging "success" messages from unimportant
in-core query handlers.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2488 5f96b256-904b-4d8d-8c98-d829582c6739