qh: Be less strict with query format
With this patch, we stop requiring '@' or '#' as query-prefixes.
They can still be used to denote "keepalive" or "oneshot" queries,
respectively, but we no longer require them to be present and fall
back to 'keepalive' if neither of them is. Queries without a magic
byte-prefix will thus get a persistent connection.

We also stop requiring a specific query and pass a nul byte buffer
with length 0 in case none is passed. Handlers can respond to this
in any way they like, although a strong suggestion (which we also
follow) is that (empty-query) = "help", unless other compelling
options exist.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2493 5f96b256-904b-4d8d-8c98-d829582c6739