xodtemplate: Fix crash when referring to invalid servicegroups
The code for printing error messages when a service is referring to an
invalid servicegroup was incorrect, because it used the wrong pointer -
which can be uninitialized - for writing file and line of the error.

In other words: when you do things incorrectly, nagios will no longer do
so too.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2494 5f96b256-904b-4d8d-8c98-d829582c6739