nerd: Make API useful for eventbroker modules
Since 'broadcast()' was made static and thus limited to the file
nerd.c and no other way existed that let callers get the list of
subscribers or communicate with them in any way, the NERD api
was quite incomplete earlier. With this patch, we instantly make
it reasonably usable.

broadcast() is renamed to nerd_broadcast() and published to the
rest of Nagios and we add nerd_get_subscriptions() as well as
nerd_get_channel_id() as publicly available functions.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2496 5f96b256-904b-4d8d-8c98-d829582c6739