base/notifications: Do send notifications when services goes down
So, this was a bit silly...

When checking a service alert's notification viability, we forgot
to check if the alert was about a recovery before ensuring that the
recovery was valid to send out. This means that people who've tried
upgrading have usually received some notifications - just enough not to
think too hard about it - and people who haven't have still received
other types of notifications.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2497 5f96b256-904b-4d8d-8c98-d829582c6739