xod: Make group member parsing work with precached objects
This caused all manner of weird bugs, such as hostgroups being
empty and notifications not being sent out to contacts added to
contactgroups which in turn were set to be assigned to certain
hosts and services, but only when using precached objects, which
is how this slipped through the testing net.

With this patch the objects.precache is identical to the
objects.cache file generated from it after Nagios has parsed it,
with two exceptions:
 * The date/time thing will obviously differ.
 * All lists are printed in reverse order in objects.cache

The date thing I won't even bother commenting on. Just live with
it. The reverse order thing might be a bit of an issue though,
but since member lists aren't sorted in either file (and won't be
if Nagios is started without using a precache file) anyway, I
really don't care about that either.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2504 5f96b256-904b-4d8d-8c98-d829582c6739