core: Make perfdata and status files nagios.cfg-relative
For most people this won't matter in the slightest, as most people
already have absolute paths for them, but it's nifty for testing
and makes path resolution consistent with the ones parsed in config.c.
Since where we parse a variable is an implementation detail that we
can't expect users to be aware of, this should probably be considered
a bugfix, although in reality it's more like a feature extension.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2508 5f96b256-904b-4d8d-8c98-d829582c6739