cgis: Provide cgi's with a 'config_file_dir' variable
This makes it possible to use nagios.cfg-relative paths for things
in the core while letting the cgi's compile properly.

What *should* be done is to unify the config parsing for both cgi's
and daemon, so they use the same code to parse the same file and
get the same variables out of it, but at least it builds now.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2513 5f96b256-904b-4d8d-8c98-d829582c6739