lib/iobroker: Avoid returning IOBROKER_ENOSET when it's clearly bogus
We already have a check for IOBROKER_ENOSET above the one where we
run epoll_ctl(), and we can never get to the epoll_ctl() call if we
have no iobroker set, so make sure we return something sensible when
epoll_ctl() fails.

This might be redundant with the earlier patch. If so, we'll fix that
later.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2518 5f96b256-904b-4d8d-8c98-d829582c6739