core: Make loadctl work again
It's funny how much headache a typo can make, and especially when
it's an arrow pointing in the wrong direction. It could, for example
mean the difference between life and death if it's a roadsign. In
this case, it wasn't quite as earthshattering as all that, but we did
manage to hog 100% of one cpu by never (ever) running an event and
iterating over the event queue almost as fast as we run out of beer
at the office. All because an arrow was pointing the wrong way.

I really should test things like that, but how fun would life be
without surprises?

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2525 5f96b256-904b-4d8d-8c98-d829582c6739