qh: Fortify removal of registered handlers
We want to remove registered handlers from the (now doubly) linked
list of registered handlers, as well as making sure that 'qhandlers'
always points to the first entry if there are any or is NULL in
case there's not.

This patch makes it so, meaning the rest of Nagios, including its
modules, can safely register handlers upon initialization without
having to double-check if the error they get back is because they're
already registered or if it's due to something more serious.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2529 5f96b256-904b-4d8d-8c98-d829582c6739