qh: Set qh_table = NULL when deinitializing the queryhandler
This ensure that we never end up in a situation where we get queries
in between deinitializing all handlers and re-enabling the handler
altogether.

I'm not sure if we really *can* end up in such a situation, but if
we can't now then it's an at least theoretical possibility in case
the linux kernel folk decide to change how epoll() works. It may
also matter for the case where we're using something other than epoll.

Anyways, this is a belts-and-suspenders-and-tidyness patch, and it
definitely won't hurt anything, so let's apply it.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2532 5f96b256-904b-4d8d-8c98-d829582c6739