core: Make debugging coredumps infinitely more pleasant
The fact that gdb can never find the files we demandload is pretty
damn annoying. With this patch we make no great architectural changes,
but we do make a double copy of the module, so that a file with its
exact contents exist at the place where we loaded it from, but with
a different inode.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2533 5f96b256-904b-4d8d-8c98-d829582c6739