nerd: Name our in-core module 'NERD' as filename
Since we normally log the filename of the module when initializing
and deinitializing modules, we might as well abuse the fact and name
our module 'NERD'. Since it's a core module, we'll never *really* try
to load that (hopefully) nonexistant file, but at least we'll stop
printing "Successfully deinitialized (null)", which just looks weird.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2534 5f96b256-904b-4d8d-8c98-d829582c6739