objects: next_check_event member of struct host is a timed_event
It's always used as that, and never as anything else. This patch
makes sure the compiler warns us if we try to do something silly
with it. Any modules that tried to use it as anything else would
have either caused Nagios to crash, or failed to do its job, so
this isn't actually an ABI change (as the object size remains
constant), but an ABI bugfix which just lets the compiler help
us more, and getting less in our face in case we forget to cast
the damned thing to the timed_event it really is.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2535 5f96b256-904b-4d8d-8c98-d829582c6739