core: Drop privileges before opening debug log
Otherwise the debug log will (most of the time) be owned by root,
and won't be amended to after the first reload. This means we have
to drop debug logging from the drop_privileges() function, but
since it only ever does anything while we're still in the foreground
that's really not a major issue.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2539 5f96b256-904b-4d8d-8c98-d829582c6739