core: Daemonize before we open any persistent files
Once upon a time there was a keepalive daemon that figured it would
be a mighty fine idea if it closed stdin, stdout and stderr for the
programs it launched. That was kinda stupid, since it caused Nagios
to fail to use its IO broker set. The same sort of thing can happen
with logfiles and other things that uses filedescriptors, of all
kinds, so we avoid opening *any* persistent files before we're done
close()'ing stdin, stdout and stderr.

This means we no longer print the "Finished daemonizing..." message
in the logfile, since it would look pretty stupid there *after* the
"Nagios starting" message (which now prints the correct pid the first
time around, btw).

It also means we never load or initialize modules in any way before
we're done daemonizing, since we'd only have to unload them if we
fail, and they could be subject to the same "oopsie, I killed your
filedescriptor" issue as Nagios core was under such exceptionally
weird and hard-to-debug circumstances.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2540 5f96b256-904b-4d8d-8c98-d829582c6739