buildsystem: Compile nagios.o before linking it
This makes it infinitely more pleasant to run all sorts of static
code analysis tools, which is good, since they sometimes catch
real errors, which are really bad. More often, they catch legacy
code that has become nonsensical by more recent patches and only
serve to confuse developers. Since we coders are so confused as
it is, we might as well avoid more of the same as much as we can.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2541 5f96b256-904b-4d8d-8c98-d829582c6739