core: Try really hard to log errors regarding logging during reloads
If the user has made a typo in the log_file path and issues a reload,
we didn't print an error message anywhere at all, but silently failed
to log anything at all.

With this patch, we'll redirect such logs to the old logfile, but we
consider it an error that we can't write to the newly configured file
and return an error code promptly.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2543 5f96b256-904b-4d8d-8c98-d829582c6739