Changed check results utility functions to accept list pointer.
In order to make the check results utility functions more generic, they have
been changed to accept a check results list pointer rather than using the 
global check_results_list. This will enable these function to be used to 
manipulate temporary lists without impacting the system's global list of
check results.

This change is being made in anticipation of implementing a function to 
see which checks are still executing at a Nagios restart so those checks
are not rescheduled incorrectly as a result of the restart.


git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/branches/nagios-3-4-x@2552 5f96b256-904b-4d8d-8c98-d829582c6739