Change ownership on debug log file
Previously, when the debug log file was opened, it was opened before Nagios
dropped privileges. If the debug log file did not exist, it would be created
by the privileged owner (root). Because the default umask is 022, it 
would be mode 644. If Nagios received a HUP signal, it close and attempt 
to reopen the file as the unprivileged user, which would fail. 

This patch changes ownership on the debug log file in the drop_privilegeds()
function once the uid and gid for the unprivileged user have been determined.


git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/branches/nagios-3-4-x@2553 5f96b256-904b-4d8d-8c98-d829582c6739