Fixed bug #408 - service checks get duplicated on reload.
If a check is still running when Nagios is restarted or reloaded (via SIGHUP),
Nagios will now look for running checks and mark them as executing. Then, when
the checks are scheduled before the execution loop starts, the running checks
will be not be rescheduled. They will be rescheduled as usual when their 
results are reaped.

This commit also re-introduces the extra event debug logging that was 
unintentionally removed between versions 3.4.2 and 3.4.3.


git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/branches/nagios-3-4-x@2554 5f96b256-904b-4d8d-8c98-d829582c6739