core: Handle removing the current event on I/O broker input
This fixes the possibly silliest bug in the entire world. If one
scheduled flexible downtime which wasn't triggered and a several
commands to remove it arrives between the stop event is peeked
from the scheduling queue but before it's actually executed, we
would end up accessing an already free()'d event, usually
resulting in a SIGSEGV or a SIGBUS.

The maximum window for that was 1.5 seconds (although usually
considerably smaller since internal housekeeping events are
scheduled with fairly frequent intervals), and both commands
had to arrive within that window for the bug to be triggered.

Like I said; Possibly the silliest bug in the world.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2578 5f96b256-904b-4d8d-8c98-d829582c6739