xodtemplate: Avoid free()'ing dependent hostlist multiple times
For the (rather unusual) case when we have hosts depending on
multiple others using hostdependencies, we used to crash with a
segfault when loading the configuration (regardless of whether we
verified it to start running or just as a part of a config restart
test thing).

The problem was due to some overexcited optimization on my behalf,
attempting to free() the variables of an inner loop while parsing
it, regardless of whether we needed to parse it again later or
not.

With this patch, the config verifies nicely and Nagios can load
it without problems.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2606 5f96b256-904b-4d8d-8c98-d829582c6739